fix(retroarch): Work around missing params on uninstall
Work around choco#1479 by saving package parameters as XML in the package
lib directory, and restoring as necessary on uninstall in order to get
info we need to properly clean up.

+++ retroarch/tools/chocolateyUninstall.ps1
@@ -7,16 +7,29 @@ Uninstall-ChocolateyZipPackage "$packageName" "$zip"
 
 $toolsPath = "$(Split-Path -parent $MyInvocation.MyCommand.Definition)"
 
+# Get package parameters
 $pp = Get-PackageParameters
+if ($null -eq $pp -or $pp.Count -eq 0) {
+  # Work around for choco#1479 https://github.com/chocolatey/choco/issues/1479
+  $paramsFile = Join-Path (Split-Path -Parent $toolsPath) 'PackageParameters.xml'
+  if (Test-Path -Path $paramsFile) {
+    Write-Debug "Loading package parameters from $paramsFile"
+    $pp = Import-Clixml -Path $paramsFile
+  } else {
+    Write-Debug "No package parameters available; shims and desktop shortcut won't be removed if they were created"
+    $pp = @{}
+  }
+}
+
 $installDir = $toolsPath
 if ($pp.InstallDir -or $pp.InstallationPath ) { $installDir = $pp.InstallDir + $pp.InstallationPath }
 
-#if ($installDir -ne $toolsPath) { # choco issue #1479
+if ($installDir -ne $toolsPath) {
   Uninstall-BinFile retroarch -path "$installDir\retroarch.exe"
   Uninstall-BinFile retroarch_debug -path "$installDir\retroarch_debug.exe"
-#}
+}
 
-#if ($pp.DesktopShortcut) { # choco issue #1479
+if ($pp.DesktopShortcut) {
   $desktop = [System.Environment]::GetFolderPath("Desktop")
   Remove-Item "$desktop\RetroArch.lnk" -ErrorAction SilentlyContinue -Force | Out-Null
-#}
+}
